Update: 勝利条件の判定関数を別ファイルに切り出した
このゲームのドメイン部分を切り出した
